Resize the window dynamically
Gesture is starting in one of the outer region, going to all other regions
except the center one and ending past the starting one. The incremented size is
multiplied by the number of region past the starting one where you lift your
finger. Issuing this gesture clockwise grows the window, counter-clockwise
shrinks it.